fix: Adding SOPS KMS coverage
