+++ test/fixtures/auth-provider-cmd/sops/main.tf
@@ -1,0 +1,6 @@
+variable "hello" {}
+
+output "hello" {
+  value = var.hello
+}
+
