+++ test/fixtures/auth-provider-cmd/sops/secrets.json
@@ -1,0 +1,22 @@
+{
+	"data": "ENC[AES256_GCM,data:+u+2NzL7vRX52H/cmY9wOOVvCXlMpECsRDIbNiM0SsJmW3xnzRoDlgOzeGmtR9+8Y2fR37Ctou/vvXBUKp6TMHlXrtKZy0WdcM0dXGAGQHTijq56RqeEqrw+FRwfUS87zMks43TSXoAwU+PlIkzd1xhAnyghgVIeuxqCYRCrr0XIZ3ytQlQBZlnj71suNOyScEiVxKFrH68NR9Rek1LsjU4FpySlXUHFjERZFKw5LJ1Z8g/XcTsFThz9EKPiKim0kA+vxl2AMYyZkwsvxh0SZIUKbRKXDdfORgARdmbKkbH2ssbM4zYgZA==,iv:OcpVr6Va6wVzAse5u3wGVbRiNsTuF+0zIZpfsreEAAo=,tag:4UjnPurkAi9PFtBGmY85wg==,type:str]",
+	"sops": {
+		"kms": [
+			{
+				"arn": "arn:aws:kms:us-east-1:087285199408:key/bd372994-d969-464a-a261-6cc850c58a92",
+				"created_at": "2024-09-26T12:19:43Z",
+				"enc": "AQICAHi+J98ytR8eNxKA5Tt6E5ILIMsUjvQf2k7quGoMjNULtgH+PP0HJ1hFlHfbNdVna1PaAAAAfjB8BgkqhkiG9w0BBwagbzBtAgEAMGgGCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMqYcRjvhJ6EZodmUWAgEQgDvWUXLyQXYgDjl9mwTNMMqz76lXdlbfxG3SASRkMFZyhpcCNVsGIV4fcbcA4NzfT98JqPo6/XVHPc4LpA==",
+				"aws_profile": ""
+			}
+		],
+		"gcp_kms": null,
+		"azure_kv": null,
+		"hc_vault": null,
+		"age": null,
+		"lastmodified": "2024-09-26T12:19:43Z",
+		"mac": "ENC[AES256_GCM,data:xqRGYCj5euTgeFHgoKVM9moAH/+dpg2CE1466sM5T+e34bCEvVqJccCLIcDDhn1XGWpU0Ga8056oX6HIShZLiTiW/8443eUdX8n+sEe91ILNnOc+BGd+a8AC2bXHD3g+rIsRKNJ+Dau6ry+FeJPjAc9g+qVFefVCa4Ab1RmHIfw=,iv:81wJEhqdwvWRh3pRoA0cI+FqJeRBD69zWC3eq9gBRak=,tag:DtYI36PIIMCQZlhyB0Mbiw==,type:str]",
+		"pgp": null,
+		"unencrypted_suffix": "_unencrypted",
+		"version": "3.9.0"
+	}
+}
