+++ test/fixtures/auth-provider-cmd/sops/terragrunt.hcl
@@ -1,0 +1,8 @@
+locals {
+  data = jsondecode(jsondecode(sops_decrypt_file("secrets.json")).data)
+}
+
+inputs = {
+  hello = local.data.hello
+}
+
