+++ test/fixtures/sops-kms/secrets.env
@@ -1,0 +1,10 @@
+DB_USER=ENC[AES256_GCM,data:5yinT3k=,iv:DP5bS4hJeC8Znc4wHCVIpkJjvdPQOmzwy0KRimvPuuY=,tag:UWl/WagEPWdB+k5j3YQpxA==,type:str]
+DB_PASSWORD=ENC[AES256_GCM,data:1/x9ozAM,iv:2KaZz1fKcZ3h5L/8XAgKHtqLDW2AeQ0mTFOxjTBFX2Q=,tag:vGqg7EJHcQhtpAQ2/7oaAA==,type:str]
+sops_kms__list_0__map_arn=arn:aws:kms:us-east-1:087285199408:key/bd372994-d969-464a-a261-6cc850c58a92
+sops_kms__list_0__map_aws_profile=
+sops_kms__list_0__map_created_at=2024-09-26T12:55:27Z
+sops_kms__list_0__map_enc=AQICAHi+J98ytR8eNxKA5Tt6E5ILIMsUjvQf2k7quGoMjNULtgEXDu/IQ2Kn9wFn+855eUHqAAAAfjB8BgkqhkiG9w0BBwagbzBtAgEAMGgGCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMl+eJ6qMUSMxmnlEdAgEQgDsmXsaJulpUVuL2YNPSwEBEgmCx20iXKMqXwiGMZPM45w3HnV/Bjqtx2yZsiwHaR5g0GrVFya9Pi69u4A==
+sops_lastmodified=2024-09-26T12:55:27Z
+sops_mac=ENC[AES256_GCM,data:0z4OU1YSM5is/qi4l0FG0P9j1CY00Db42bR2We9iVbEbZOwRFv8XbLKxPgquVr/d756uVDUXUHU67YXZrd1z82Se0RRy6ERfOH62V0Zya/KV26Iku7k+EyZ/piHgnbpIhT0xhlaVqT6mFGphG/pdVELGwkE1l46o7xoaIq/77n4=,iv:6eamZcbm++PZU0HrflHDjsP3aqdr53g/GfNT5BDaSyo=,tag:1mOYIGgVoZOFxeZ5T2oR8A==,type:str]
+sops_unencrypted_suffix=_unencrypted
+sops_version=3.9.0
