+++ test/fixtures/sops-kms/secrets.ini
@@ -1,0 +1,13 @@
+[terragrunt]
+user     = ENC[AES256_GCM,data:v7hX9P4=,iv:k+sFBEDdUjWL5BNm9Qhkwr+GMiwmnZiKktPCiUh3oYU=,tag:SjIp1kc977n0qFPZdiH84w==,type:str]
+password = ENC[AES256_GCM,data:aASx6oEW,iv:SvW2FnbBa+nbJRvmzYsPFO5vi9AuM3W9y9OMjmJkAyc=,tag:MptmHHnwggP9eAP/89i59A==,type:str]
+
+[sops]
+lastmodified                 = 2024-09-26T12:55:27Z
+kms__list_0__map_arn         = arn:aws:kms:us-east-1:087285199408:key/bd372994-d969-464a-a261-6cc850c58a92
+kms__list_0__map_created_at  = 2024-09-26T12:55:27Z
+mac                          = ENC[AES256_GCM,data:Mf0pwX+ClEmMiG3BKgeTqVmOtwqjAMUMuVEwGfIDC7cKG5ZujVs0VDIX5vxtCildeqMC77EICmkL7ajHIYFOQBGkuks5Bd4fZ+nWg85xKBTFTf+037db9o8uKVDtlpH6za9tWC5m7O36QZsQ7yRbk6iZh8Xui5k+N6R3CTDpkis=,iv:Jf9VN/bTcC8LwWXe4cs1Nx4iRGM/zPs9JdZbmWu0IVg=,tag:lKJuo/EpI8GKToeS+enJiQ==,type:str]
+version                      = 3.9.0
+kms__list_0__map_enc         = AQICAHi+J98ytR8eNxKA5Tt6E5ILIMsUjvQf2k7quGoMjNULtgG6SyRaS5/aqtE4oneGrnRMAAAAfjB8BgkqhkiG9w0BBwagbzBtAgEAMGgGCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMioe9eMXjGoLW/NWLAgEQgDv82HUHtC/zXxkbdkaEOXf97p4MXHUCTGb9pBPFSQePYM49Y/eq0eMxgpuNMddqKlbV8S62elrk/d3qig==
+kms__list_0__map_aws_profile = 
+unencrypted_suffix           = _unencrypted
