+++ test/fixtures/sops-kms/secrets.json
@@ -1,0 +1,32 @@
+{
+	"hello": "ENC[AES256_GCM,data:chBszAAAR0LJsYe4bEm4cjbG+e+sdTS9pHdTwQ6mrH1ohbjsFidUJRExtKMKAw==,iv:WWw7Luz0OX9JrE8HUVcZzhko1mtp6NZPGi8dId8KaUM=,tag:DE7KwBkrrpaYFYSsZZLasw==,type:str]",
+	"example_key": "ENC[AES256_GCM,data:xbQTqmd9HGE5z6JaHw==,iv:0xn2eM0rEmd0oZyuJNzvEXsa6NSHYMueOkTUvahD90Y=,tag:WIqU6W9FC4d/zoJP/COhVQ==,type:str]",
+	"example_array": [
+		"ENC[AES256_GCM,data:sF/AnbmGaH4rqPASEhU=,iv:5aywEQePONgZkwwavOssMY2s+WDrswuODxLl2lLMSks=,tag:EnGDzBaiUQ+2DybAu4sCfg==,type:str]",
+		"ENC[AES256_GCM,data:iSmgNloOVDQMjLlG7FY=,iv:sKv9zK38XRT9VFBbKWqDHmzOhXh9w53wa5u2RFXQwYk=,tag:RXIjEQ4kE1CjwWuT7tV0JQ==,type:str]"
+	],
+	"example_number": "ENC[AES256_GCM,data:K/1D8tDBNNOapA==,iv:5Djbqi7oyZ4nTFpJvMlMYBIYWwRQ3VCRS0JAu103u6g=,tag:b+WfoiHTwgoX4CkoNOIf/w==,type:float]",
+	"example_booleans": [
+		"ENC[AES256_GCM,data:6bIc5A==,iv:02H7C/7RlwXpOguEy+IeLpJxxY8W8K4zKHT9Ivhniqo=,tag:f6sCenNwOJjcuVeDj/V7Pg==,type:bool]",
+		"ENC[AES256_GCM,data:0x8KKeo=,iv:8AK/o4fuq5daQJDRz7qonbd7ZEpxWHcaT2fNCizbGI0=,tag:OmXXNL7poYITcc3g4f926g==,type:bool]"
+	],
+	"sops": {
+		"kms": [
+			{
+				"arn": "arn:aws:kms:us-east-1:087285199408:key/bd372994-d969-464a-a261-6cc850c58a92",
+				"created_at": "2024-09-26T12:55:27Z",
+				"enc": "AQICAHi+J98ytR8eNxKA5Tt6E5ILIMsUjvQf2k7quGoMjNULtgFBAq5l6JQ7H6GTZ0s6/k8nAAAAfjB8BgkqhkiG9w0BBwagbzBtAgEAMGgGCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMZ7I5xuFWUUyrSD2hAgEQgDt6p+gM9IpPQo1adat2MfDHbdigWDZW1oUSIt+XmvLSzU8stPsjSBVlgdMIncnCaEp/28eT2+fJOOHPhw==",
+				"aws_profile": ""
+			}
+		],
+		"gcp_kms": null,
+		"azure_kv": null,
+		"hc_vault": null,
+		"age": null,
+		"lastmodified": "2024-09-26T12:55:28Z",
+		"mac": "ENC[AES256_GCM,data:sqV9clMyATtpp26a1xReJwZMoCrn9Ee/XETXzP/1flnVQ0RHXqbQYLd0zWIppK/eaXpYkZiwhQ6mpzuiDDTA53P4K4qMdpAt1zQ1enw2tAggDMD3O9zJ5ZvyXaYhhZKxYuUOWmwzchaQXTxvFClcAfCDA5MyrFNylHd7vV06ERY=,iv:VO2RgVFRDa7KqKZC91Mmo4JWKqz0hI8/JRG3sHBEh3I=,tag:FKyzL5oCQdacmOaPRz0gmg==,type:str]",
+		"pgp": null,
+		"unencrypted_suffix": "_unencrypted",
+		"version": "3.9.0"
+	}
+}
