+++ test/fixtures/sops-kms/secrets.txt
@@ -1,0 +1,22 @@
+{
+	"data": "ENC[AES256_GCM,data:gbo1A3QA/JGY49OwjwESxxlA,iv:hZe6bia9/114NsW++jWAxis2fpHdQ+CUJlgrUVD9n88=,tag:z6fo5qFE82otY4JTBWzFPQ==,type:str]",
+	"sops": {
+		"kms": [
+			{
+				"arn": "arn:aws:kms:us-east-1:087285199408:key/bd372994-d969-464a-a261-6cc850c58a92",
+				"created_at": "2024-09-26T12:55:28Z",
+				"enc": "AQICAHi+J98ytR8eNxKA5Tt6E5ILIMsUjvQf2k7quGoMjNULtgGeVuQinJDidP/+IEwoH4qbAAAAfjB8BgkqhkiG9w0BBwagbzBtAgEAMGgGCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMCSYnXMXk1Htl2232AgEQgDvSL7+snq0vI68TvmzvYQllB/j77Fhd/i5I1oVfk0Hg9n7iUDcqFt65HHx6cuy6qLCf6ZZrWMu/VvtwJw==",
+				"aws_profile": ""
+			}
+		],
+		"gcp_kms": null,
+		"azure_kv": null,
+		"hc_vault": null,
+		"age": null,
+		"lastmodified": "2024-09-26T12:55:28Z",
+		"mac": "ENC[AES256_GCM,data:Da2QYj51vS0BRvAwA8wQWOYgS9clOeKlPxSKXdnM8up04mFfnSl1GDC8uGrHa3ZurCjjzK4mi9Ug/0uXt9cNFTPu9WWBVpmK2NGUjB5j1dfwwFPI4LWwICogYPcODE/RtIy0ROhHfNa4ZRpgEwUyGikFqtG1gWjdaPT4sdpu0u0=,iv:qb17WZHOfWZLdiJQAzYHr9PY8XKccsizNTriw7KQI3o=,tag:g+NaMWPpG8fOvIA4f0EejQ==,type:str]",
+		"pgp": null,
+		"unencrypted_suffix": "_unencrypted",
+		"version": "3.9.0"
+	}
+}
