+++ test/fixtures/sops-kms/secrets.yaml
@@ -1,0 +1,25 @@
+hello: ENC[AES256_GCM,data:UctdQHxu6o2NCEM4CtRZDYNDQM2FSG0kbz4ZiTBrD1dQu9gHOQHqVyvVffaC/A==,iv:txyqGoTHwiM07kGIa0WQvMLZsicLpavYfbLCWQZZ/ME=,tag:VhGqTrCBievJIe0Xbbuzzg==,type:str]
+example_key: ENC[AES256_GCM,data:5UeifFtzkkfUF72Fyg==,iv:heBSgumZKtl9SEBI8hxLMXRGy9TAQAWuAwOr9zI75d4=,tag:+ePgvZ46ZpvjRarvNO+acw==,type:str]
+#ENC[AES256_GCM,data:uK3RAvakhDn/GXfblb7iiQ==,iv:hHr/W4qRoON+B1gWHIJL5yRbYkFJ+n5IiG6Zzd3WnOU=,tag:E9Tv6j85gFCjjVlAW4AxIw==,type:comment]
+example_array:
+    - ENC[AES256_GCM,data:vqHuKNctFSMHgqjwlEM=,iv:c4lP+AhlyMLWibKy9CponrUUMOfIO4o5gDIoKp+Nx4Y=,tag:d29u2zfxxbIlYr+73rqWZQ==,type:str]
+    - ENC[AES256_GCM,data:uJwuLNoKYY9v40Kq3LY=,iv:guNMIvtzpUimYHsnqZ20WYsEVhWpQhQbh3ggw6uSIzA=,tag:V32h8FtuDUnlBZI6JYkQgw==,type:str]
+example_number: ENC[AES256_GCM,data:3yc0r2aSfHxy,iv:3wrKIene9z0q4NuS/Nc3SqZG506tpFETbaTGtUUoQkc=,tag:EwBNV3sBFQev4aukf20XTw==,type:float]
+example_booleans:
+    - ENC[AES256_GCM,data:kqRt/A==,iv:qssmJd6Of8tfTpIozUzXu5+xcJuEMNWa6W5I7vOaPME=,tag:3hUar1Yp4/eoq2Z75EdidA==,type:bool]
+    - ENC[AES256_GCM,data:GKuXxS8=,iv:rHeGQpu/GPwykvlQebE+nliyMJQsLrmiCG0HCt06A7c=,tag:Jb3H/HTkB0NaWLZsXSpDzQ==,type:bool]
+sops:
+    kms:
+        - arn: arn:aws:kms:us-east-1:087285199408:key/bd372994-d969-464a-a261-6cc850c58a92
+          created_at: "2024-09-26T12:55:28Z"
+          enc: AQICAHi+J98ytR8eNxKA5Tt6E5ILIMsUjvQf2k7quGoMjNULtgHK/dxuabG3Ua5XjTrtqSBQAAAAfjB8BgkqhkiG9w0BBwagbzBtAgEAMGgGCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQM4MCxCLTdfXZmRg2WAgEQgDufPkOmTiJyNkD+k3o75tJAwAUz/rDuADkGT3GP3Gs8I6BIzoQDkTJO7RgCcq6GMi3ClEtsflvgpP+WqQ==
+          aws_profile: ""
+    gcp_kms: []
+    azure_kv: []
+    hc_vault: []
+    age: []
+    lastmodified: "2024-09-26T12:55:28Z"
+    mac: ENC[AES256_GCM,data:cZaMv1T0LEWA+N40FIiZrTTIC18/XkODrFmoK81+fb8CyxxJB2iciTgS0lRlL0imJ1uo+RfZ4+OKSK9O4L2x4RR9lGte4y7XMvb6g1+zIB5VH8SxNnQzJZLR2xlVdMsrqQu7F7LypKGuD07Rhk+exRlyF+opHjY551SEREpIftU=,iv:xH+iSvF1Akp8bkuxIpiLd8TUPE+1IBKPvAMnSlHofIQ=,tag:9Nl4U7+rJ0IFEhfQYhfIDQ==,type:str]
+    pgp: []
+    unencrypted_suffix: _unencrypted
+    version: 3.9.0
