+++ test/integration_test.go
@@ -98,8 +98,6 @@ const (
 	testFixtureRefSource                      = "fixtures/download/remote-ref"
 	testFixtureSkip                           = "fixtures/skip/"
 	testFixtureSkipDependencies               = "fixtures/skip-dependencies"
-	testFixtureSops                           = "fixtures/sops"
-	testFixtureSopsErrors                     = "fixtures/sops-errors"
 	testFixtureSourceMapSlashes               = "fixtures/source-map/slashes-in-ref"
 	testFixtureStack                          = "fixtures/stack/"
 	testFixtureStdout                         = "fixtures/download/stdout-test"
@@ -2960,73 +2958,6 @@ func runValidateAllWithIncludeAndGetIncludedModules(t *testing.T, rootModulePath
 	return includedModules
 }
 
-// sops decrypting for inputs
-func TestSopsDecryptedCorrectly(t *testing.T) {
-	t.Parallel()
-
-	cleanupTerraformFolder(t, testFixtureSops)
-	tmpEnvPath := copyEnvironment(t, testFixtureSops)
-	rootPath := util.JoinPath(tmpEnvPath, testFixtureSops)
-
-	runTerragrunt(t, "terragrunt apply -auto-approve --terragrunt-non-interactive --terragrunt-working-dir "+rootPath)
-
-	stdout := bytes.Buffer{}
-	stderr := bytes.Buffer{}
-
-	err := runTerragruntCommand(t, "terragrunt output -no-color -json --terragrunt-non-interactive --terragrunt-working-dir "+rootPath, &stdout, &stderr)
-	require.NoError(t, err)
-
-	outputs := map[string]TerraformOutput{}
-	require.NoError(t, json.Unmarshal(stdout.Bytes(), &outputs))
-
-	assert.Equal(t, []interface{}{true, false}, outputs["json_bool_array"].Value)
-	assert.Equal(t, []interface{}{"example_value1", "example_value2"}, outputs["json_string_array"].Value)
-	assert.InEpsilon(t, 1234.56789, outputs["json_number"].Value, 0.0001)
-	assert.Equal(t, "example_value", outputs["json_string"].Value)
-	assert.Equal(t, "Welcome to SOPS! Edit this file as you please!", outputs["json_hello"].Value)
-	assert.Equal(t, []interface{}{true, false}, outputs["yaml_bool_array"].Value)
-	assert.Equal(t, []interface{}{"example_value1", "example_value2"}, outputs["yaml_string_array"].Value)
-	assert.InEpsilon(t, 1234.5679, outputs["yaml_number"].Value, 0.0001)
-	assert.Equal(t, "example_value", outputs["yaml_string"].Value)
-	assert.Equal(t, "Welcome to SOPS! Edit this file as you please!", outputs["yaml_hello"].Value)
-	assert.Equal(t, "Raw Secret Example", outputs["text_value"].Value)
-	assert.Contains(t, outputs["env_value"].Value, "DB_PASSWORD=tomato")
-	assert.Contains(t, outputs["ini_value"].Value, "password = potato")
-}
-
-func TestSopsDecryptedCorrectlyRunAll(t *testing.T) {
-	t.Parallel()
-
-	cleanupTerraformFolder(t, testFixtureSops)
-	tmpEnvPath := copyEnvironment(t, testFixtureSops)
-	rootPath := util.JoinPath(tmpEnvPath, testFixtureSops)
-
-	runTerragrunt(t, fmt.Sprintf("terragrunt run-all apply -auto-approve --terragrunt-non-interactive --terragrunt-working-dir %s/../.. --terragrunt-include-dir %s", rootPath, testFixtureSops))
-
-	stdout := bytes.Buffer{}
-	stderr := bytes.Buffer{}
-
-	err := runTerragruntCommand(t, fmt.Sprintf("terragrunt run-all output -no-color -json --terragrunt-non-interactive --terragrunt-working-dir %s/../.. --terragrunt-include-dir %s", rootPath, testFixtureSops), &stdout, &stderr)
-	require.NoError(t, err)
-
-	outputs := map[string]TerraformOutput{}
-	require.NoError(t, json.Unmarshal(stdout.Bytes(), &outputs))
-
-	assert.Equal(t, []interface{}{true, false}, outputs["json_bool_array"].Value)
-	assert.Equal(t, []interface{}{"example_value1", "example_value2"}, outputs["json_string_array"].Value)
-	assert.InEpsilon(t, 1234.56789, outputs["json_number"].Value, 0.0001)
-	assert.Equal(t, "example_value", outputs["json_string"].Value)
-	assert.Equal(t, "Welcome to SOPS! Edit this file as you please!", outputs["json_hello"].Value)
-	assert.Equal(t, []interface{}{true, false}, outputs["yaml_bool_array"].Value)
-	assert.Equal(t, []interface{}{"example_value1", "example_value2"}, outputs["yaml_string_array"].Value)
-	assert.InEpsilon(t, 1234.5679, outputs["yaml_number"].Value, 0.0001)
-	assert.Equal(t, "example_value", outputs["yaml_string"].Value)
-	assert.Equal(t, "Welcome to SOPS! Edit this file as you please!", outputs["yaml_hello"].Value)
-	assert.Equal(t, "Raw Secret Example", outputs["text_value"].Value)
-	assert.Contains(t, outputs["env_value"].Value, "DB_PASSWORD=tomato")
-	assert.Contains(t, outputs["ini_value"].Value, "password = potato")
-}
-
 func TestShowErrorWhenRunAllInvokedWithoutArguments(t *testing.T) {
 	t.Parallel()
 
@@ -3774,22 +3705,6 @@ func TestTerragruntRunAllPlanAndShow(t *testing.T) {
 
 	// Verify that output contains the plan and not just the actual state output
 	assert.Contains(t, stdout, "No changes. Your infrastructure matches the configuration.")
-}
-
-func TestTerragruntLogSopsErrors(t *testing.T) {
-	t.Parallel()
-
-	// create temporary directory for plan files
-	tmpEnvPath := copyEnvironment(t, testFixtureSopsErrors)
-	cleanupTerraformFolder(t, tmpEnvPath)
-	testPath := util.JoinPath(tmpEnvPath, testFixtureSopsErrors)
-
-	// apply and check for errors
-	_, errorOut, err := runTerragruntCommandWithOutput(t, "terragrunt apply --terragrunt-non-interactive --terragrunt-log-level debug --terragrunt-working-dir "+testPath)
-	require.Error(t, err)
-
-	assert.Contains(t, errorOut, "error decrypting key: [error decrypting key")
-	assert.Contains(t, errorOut, "error base64-decoding encrypted data key: illegal base64 data at input byte")
 }
 
 func createTmpTerragruntConfigWithParentAndChild(t *testing.T, parentPath string, childRelPath string, s3BucketName string, parentConfigFileName string, childConfigFileName string) string {
